... and effect state, too!
It will be important in future to make sure that the effect state is stable. If the effect state's layout needs to change, it should be done such that it accepts previous layouts, and previous versions do something similar if possible.

Reading can be sorted by designating one of the fields as "version" and doing some re-shuffling to convert on read, but I'm not sure about how to write them.  Maybe that produces an error, prompting you to update the effect?