Separate text input into control_textinput()
Because the UI system is stateless, you have to keep the state yourself and pass it in every time.  The state is two integer values (the select range, or -1 for unfocused) packed into a single floating-point value.

Because that packing is length-dependent, current selection behaviour is slightly odd if the string changes under your feet, but it's good enough for now. :)